pyproject.toml
[project]
name = "url-shortener"
version = "0.1.0"
description = "URL Shortener"
authors = [
    {name = "Satoru",email = "satoru@example.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "flask (>=3.1.3,<4.0.0)",
    "flask-bcrypt (>=1.0.1,<2.0.0)",
    "flask-limiter (>=4.1.1,<5.0.0)",
    "flask-login (>=0.6.3,<0.7.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "gunicorn (>=25.1.0,<26.0.0)",
    "user-agents (>=2.2.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
